Allow insert row on table with no rows. SQL syntax highlighting on query tabs. Format changes.
